[Test] Revert changes to SourceKit documentation tests
