+++ test/SourceKit/CodeComplete/complete_from_clang_module.swift
@@ -1,11 +1,17 @@
 import Foo
 
-#^COMPLETE^#
-
 // REQUIRES: objc_interop
-// RUN: %target-swift-ide-test -code-completion -source-filename %s -F %S/../Inputs/libIDE-mock-sdk -code-completion-token=COMPLETE -code-completion-comments=true | %FileCheck %s -check-prefix=CHECK
+// RUN: %sourcekitd-test -req=complete -pos=2:1 %s -- -F %S/../Inputs/libIDE-mock-sdk %s | %FileCheck %s
 
-// CHECK:            Decl[GlobalVar]/OtherModule[Foo]:   fooIntVar[#Int32#];
-// CHECK-SAME-LABEL: name=fooIntVar;
-// CHECK-SAME:       briefcomment=Aaa. fooIntVar. Bbb.;
-// CHECK-SAME:       fullcomment=<Variable file="{{.*}}" line="63" column="12"><Name>fooIntVar</Name><USR>c:@fooIntVar</USR><Declaration>var fooIntVar: Int32</Declaration><Abstract><Para> Aaa.  fooIntVar.  Bbb.</Para></Abstract></Variable>
+// CHECK-LABEL:      key.name: "fooIntVar",
+// CHECK-NEXT:       key.description: "fooIntVar",
+// CHECK-NEXT:       key.typename: "Int32",
+// CHECK-NEXT:       key.doc.brief: "Aaa.  fooIntVar.  Bbb.",
+// CHECK-NEXT:       key.context: source.codecompletion.context.othermodule,
+// CHECK-NEXT:       key.typerelation: source.codecompletion.typerelation.unknown,
+// CHECK-NEXT:       key.num_bytes_to_erase: 0,
+// CHECK-NEXT:       key.associated_usrs: "c:@fooIntVar",
+// CHECK-NEXT:       key.modulename: "Foo",
+// CHECK-NEXT:       key.sourcetext: "fooIntVar"
+// CHECK-NEXT:     },
+
